Clean up code to try different file extensions.
This shortens the previous code by using stdlib functions for string manipulation here. It
also checks that the buffer is actually large enough to support the added new content.

I have not tested this yet, so I'm not making a pull request yet. It's just something that I
saw that bothered me.